requirements.txt
mw-api-client==3.2.1
requests==2.25.1
discord.py==1.6.0
pylint==2.9.3
